# chore(wallet) optimize activity filter query speed