… VERSION
… appdatabase/migrations/bindata.go