### Updates status-desktop [#11351](https://github.com/status-im/status-desktop/issues/11351)

Initial benchmarks are in place but got only minor optimizations. The benchmarks revealed some more performance considerations that will be addressed with further PRs.

- [x] Build benchmark and test all columns improvements
- [x] test all columns improvements
- [x] test alternative columns that might not help much

Also

- refactor tests as needed
- don't filter temporary or incomplete transfers (transfers.loaded != 1)

###  Benchmark results

Transaction count `100k`, benchtime=`100x`

| Test Case | ns/op (Original) | Milliseconds (Original) | ns/op (Index Timestamp) | Speed Increase (Timestamp) | ns/op (Index Addresses) | Speed Increase (Addresses) | ns/op (Index Loaded) | Speed Increase (Loaded) | ns/op (All Indexed) | Speed Increase (All Indexed) | ns/op (MT ID) | Speed Increase (MT ID)|
| --------- | ---------------- | ----------------------- | ----------------------- | ---------------------- | --------------------- | ---------------------- | ------------------ | ------------------ | ----------------- | ---------------------- | ----------------- | ---------------------- |
| RAM_NoFilter | 94730296 | 94.73 | 24182056 | 3.92 | 24428485 | 0.99 | 26649655 | 0.91 | 36456950 | 0.73 | 25841944 | 1.41 |
| SSD_NoFilter | 94674641 | 94.67 | 24201765 | 3.91 | 24335566 | 0.99 | 26726933 | 0.91 | 36442376 | 0.73 | 25801195 | 1.41 |
| SSD_AllAddresses | 108113453 | 108.11 | 24328999 | 4.45 | 24762574 | 0.98 | 26759620 | 0.92 | 41373046 | 0.65 | 30897776 | 1.33 |
| SSD_AllAddresses_AllTos | 61154882 | 61.15 | 61451549 | 0.995 | 61414241 | 1.00 | 62727450 | 0.98 | 73841549 | 0.85 | 50019852 | 1.47 |
| SSD_OneAddress | 64560310 | 64.56 | 24560320 | 2.63 | 24304115 | 1.01 | 26672523 | 0.91 | 37063820 | 0.72 | 26586419 | 1.39 |
| SSD_MovingWindow            | -                 | -                       | -                          | -                       | -                          | -                   | -                       | -         | 36469893                    | - | 26695349 | 1.36 |

Note: The "Speed Increase" column indicates how many times faster the newly indexed brings

### Demo

Filtering on 10K+ entries

Before

https://github.com/status-im/status-go/assets/47554641/04fec8ce-0535-457d-be08-04427930fd1c

After

https://github.com/status-im/status-go/assets/47554641/8fbae73a-0224-4518-b103-6e8063fd9a69