@status-im-auto 2023-07-12
### Jenkins Builds
| :grey_question: | Commit | :hash: | Finished (UTC) | Duration | Platform | Result |
|-|-|-|-|-|-|-|
| :heavy_check_mark: | 602df306 | [#1](https://ci.status.im/job/status-go/job/prs/job/linux/job/PR-3744/1/) | 2023-07-12 12:02:36 | ~3 min | `linux` | [:package:`zip`](https://status-im-prs.ams3.cdn.digitaloceanspaces.com/status-go-desktop-230712-115922-602df3-pr3744.zip) |